Add UI system with retro styling and referee speech bubbles
Major additions:
- Implemented SpriteNumberDisplay for retro-style score display
- Added RefereeSpeechBubble component for match announcements
- Created HealthCircles system with color-coded health states
- Added RefereeManager with speech bubble functionality

UI Enhancements:
- Updated fonts: Commodore Pixelized v1.2 and VT323
- Created PixelCircle sprite for health indicators
- Added SpeechBubble sprite for referee announcements
- Positioned all UI elements (scores, health, labels, timer)

Scene & Settings:
- Updated SampleScene with complete UI layout
- Modified URP rendering settings
- Added editor tools for sprite creation (CreatePixelCircleSprite)
- Added SafeMCPOperations for Unity MCP integration

Assets:
- Added 13 development screenshots documenting UI alignment
- Updated Sensei sprite
- Cleaned up old Commodore-64 font

Code improvements:
- Enhanced MatchManager score tracking system
- Fixed SpriteNumberDisplay alignment and scaling
- Added comprehensive XML documentation

+++ Assets/Scripts/UI/UIManager.cs
@@ -61,7 +61,7 @@ namespace InternationalKarate.UI
         /// <summary>
         /// Update player score display
         /// </summary>
-        public void UpdateScore(int playerNumber, float score)
+        public void UpdateScore(int playerNumber, int score)
         {
             if (playerNumber == 1 && player1ScoreDisplay != null)
             {
